# packages/contracts/src/homebrew_mlflow/contracts/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "Homebrew MLflow API",
    "version": "0.1.0"
  },
  "servers": [{"url": "/"}],
  "paths": {
...
    "/api/v1/projects/{project_id}/runs": {
      "get": {
        "operationId": "listProjectRuns", "summary": "List canonical Runs", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Runs", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/RunResponse"}}}}}}
      },
      "post": {
        "operationId": "createRun", "summary": "Create and start a locally executed Run",
        "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateRunRequest"}}}},
        "responses": {"200": {"description": "Run started", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunResponse"}}}}, "403": {"description": "Contributor required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
...
    "/api/v1/projects/{project_id}/secret-context": {
      "get": {
        "operationId": "getProjectSecretContext", "summary": "Get safe Infisical routing and reconciliation state", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Non-secret Infisical mapping", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SecretContextResponse"}}}}}
      },
      "put": {
        "operationId": "configureProjectSecretContext", "summary": "Configure non-secret Infisical routing", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ConfigureSecretContextRequest"}}}},
        "responses": {"200": {"description": "Queued reconciliation", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SecretContextResponse"}}}}, "403": {"description": "Maintainer required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
...
    "/api/v1/projects/{project_id}/experiments": {
      "get": {
        "operationId": "listProjectExperiments", "summary": "List canonical Experiments", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "include_archived", "in": "query", "required": false, "schema": {"type": "boolean", "default": false}}],
        "responses": {"200": {"description": "Experiments", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/ExperimentResponse"}}}}}}
      }
    },
    "/api/v1/projects/{project_id}/experiments/{experiment_id}": {
      "delete": {
        "operationId": "archiveExperiment", "summary": "Archive an Experiment without deleting Run history", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "project_id", "in": "path", "required": true, "schema": {"type": "string"}}, {"name": "experiment_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"204": {"description": "Experiment archived"}, "403": {"description": "Maintainer required", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "getRunDetail", "summary": "Get Run metadata, metric history, and immutable provenance", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Run detail", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunDetailResponse"}}}}, "404": {"description": "Run not visible", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
    "/api/v1/runs/{run_id}/heartbeat": {
      "post": {
        "operationId": "heartbeatRun", "summary": "Record a local Run heartbeat", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Heartbeat recorded", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunResponse"}}}}}
      }
    },
    "/api/v1/runs/{run_id}/finalize": {
      "post": {
        "operationId": "finalizeRun", "summary": "Finalize a local Run idempotently", "security": [{"bearerAuth": []}],
        "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/FinalizeRunRequest"}}}},
        "responses": {"200": {"description": "Run finalized", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RunResponse"}}}}, "409": {"description": "Different finalization evidence", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorEnvelope"}}}}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "JWT"}
    },
    "schemas": {
...
      "ConfigureSecretContextRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["infisical_project_id", "environment_slug"],
        "properties": {"infisical_project_id": {"type": "string", "maxLength": 200}, "environment_slug": {"type": "string", "maxLength": 100}, "secret_path": {"type": "string", "default": "/", "maxLength": 1000}}
      },
...
      "SecretContextResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["project_id", "infisical_project_id", "environment_slug", "secret_path", "reconciliation_state", "last_error_code", "updated_at"],
        "properties": {"project_id": {"type": "string"}, "infisical_project_id": {"type": "string"}, "environment_slug": {"type": "string"}, "secret_path": {"type": "string"}, "reconciliation_state": {"type": "string", "enum": ["queued", "in_sync", "drift", "failed"]}, "last_error_code": {"type": ["string", "null"]}, "updated_at": {"type": "string", "format": "date-time"}}
      },
...
      "CreateRunRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["repository_id", "experiment_name", "command"],
        "properties": {
          "repository_id": {"type": "string", "pattern": "^repo_[0-9A-HJKMNP-TV-Z]{26}$"},
          "experiment_name": {"type": "string", "minLength": 1, "maxLength": 200},
          "command": {"type": "array", "minItems": 1, "maxItems": 1000, "items": {"type": "string"}},
          "retry_of_run_id": {"type": ["string", "null"], "pattern": "^run_[0-9A-HJKMNP-TV-Z]{26}$"},
          "pipeline_version_id": {"type": ["string", "null"], "pattern": "^pv_[0-9A-HJKMNP-TV-Z]{26}$"},
          "environment_specification_id": {"type": ["string", "null"], "pattern": "^env_[0-9A-HJKMNP-TV-Z]{26}$"}
        }
      },
      "FinalizeRunRequest": {
        "type": "object", "additionalProperties": false,
        "required": ["exit_code", "status"],
        "properties": {
          "exit_code": {"type": "integer"},
          "status": {"type": "string", "enum": ["succeeded", "failed", "interrupted"]},
          "git_commit_sha": {"type": ["string", "null"], "pattern": "^[0-9a-f]{40,64}$"},
          "evidence": {"type": "object", "additionalProperties": true}
        }
      },
      "RunResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "project_id", "experiment_id", "repository_id", "pipeline_version_id", "environment_specification_id", "state", "command", "heartbeat_at", "ended_at", "exit_code", "logging_token"],
        "properties": {
          "id": {"type": "string", "pattern": "^run_[0-9A-HJKMNP-TV-Z]{26}$"},
          "project_id": {"type": "string"}, "experiment_id": {"type": "string"}, "repository_id": {"type": "string"},
          "pipeline_version_id": {"type": ["string", "null"]},
          "environment_specification_id": {"type": ["string", "null"]},
          "state": {"type": "string"}, "command": {"type": "array", "items": {"type": "string"}},
          "heartbeat_at": {"type": ["string", "null"], "format": "date-time"},
          "ended_at": {"type": ["string", "null"], "format": "date-time"},
          "exit_code": {"type": ["integer", "null"]},
          "logging_token": {"type": ["string", "null"], "description": "Returned only when a Run is created; bound to this Run and the tracking audience."}
        }
      },
      "ExperimentResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["id", "project_id", "name", "created_at", "archived_at"],
        "properties": {"id": {"type": "string"}, "project_id": {"type": "string"}, "name": {"type": "string"}, "created_at": {"type": "string", "format": "date-time"}, "archived_at": {"type": ["string", "null"], "format": "date-time"}}
      },
      "RunDetailResponse": {
        "type": "object", "additionalProperties": false,
        "required": ["run", "created_at", "started_at", "git_commit_sha", "finalization_evidence", "input_artifact_version_ids", "output_artifact_version_ids", "parameters", "metrics", "tags"],
        "properties": {
          "run": {"$ref": "#/components/schemas/RunResponse"},
          "created_at": {"type": "string", "format": "date-time"}, "started_at": {"type": ["string", "null"], "format": "date-time"},
          "git_commit_sha": {"type": ["string", "null"]}, "finalization_evidence": {"type": ["object", "null"]},
          "input_artifact_version_ids": {"type": "array", "items": {"type": "string"}},
          "output_artifact_version_ids": {"type": "array", "items": {"type": "string"}},
          "parameters": {"type": "array", "items": {"$ref": "#/components/schemas/ParameterInput"}},
          "metrics": {"type": "array", "items": {"$ref": "#/components/schemas/MetricInput"}},
          "tags": {"type": "array", "items": {"$ref": "#/components/schemas/TagInput"}}
        }
      },
...
      "ParameterInput": {
        "type": "object", "additionalProperties": false, "required": ["key", "value"],
        "properties": {"key": {"type": "string", "minLength": 1, "maxLength": 250}, "value": {"type": "string", "maxLength": 6000}}
      },
      "MetricInput": {
        "type": "object", "additionalProperties": false, "required": ["key", "value", "timestamp_ms"],
        "properties": {"key": {"type": "string", "minLength": 1, "maxLength": 250}, "value": {"type": "number"}, "timestamp_ms": {"type": "integer", "minimum": 0}, "step": {"type": "integer", "minimum": 0, "default": 0}}
      },
      "TagInput": {
        "type": "object", "additionalProperties": false, "required": ["key", "value"],
        "properties": {"key": {"type": "string", "minLength": 1, "maxLength": 250}, "value": {"type": "string", "maxLength": 8000}}
      },
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message", "request_id", "details", "retryable"],
        "properties": {
          "code": {"type": "string"},
          "message": {"type": "string"},
          "request_id": {"type": "string"},
          "details": {"type": "object"},
          "retryable": {"type": "boolean"}
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {"error": {"$ref": "#/components/schemas/Error"}}
      }
    }
  }
}